# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Arc House chapter banner API. Renders personalised papercut X headers server-side, pixel-identical to the in-browser preview. Nothing is persisted and photos are never stored or logged.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Which studio field the error belongs to, when it maps to one
 */
export type ApiErrorField = typeof ApiErrorField[keyof typeof ApiErrorField];


export const ApiErrorField = {
  country: 'country',
  city: 'city',
  handle: 'handle',
  role: 'role',
  photo: 'photo',
} as const;

export interface ApiError {
  /** Human-readable message safe to show inline in the studio */
  error: string;
  /** Which studio field the error belongs to, when it maps to one */
  field?: ApiErrorField;
}

/**
 * Country chapter slug; selects the papercut base and paper palette
 */
export type BannerCountry = typeof BannerCountry[keyof typeof BannerCountry];


export const BannerCountry = {
  india: 'india',
  argentina: 'argentina',
  bangladesh: 'bangladesh',
  brazil: 'brazil',
  vietnam: 'vietnam',
  france: 'france',
  nigeria: 'nigeria',
  portugal: 'portugal',
  turkey: 'turkey',
  singapore: 'singapore',
} as const;

export interface BannerRenderInput {
  country: BannerCountry;
  /**
     * City drawn (uppercased) under the chapter title. Letters, digits, spaces and . ' & / ( ) -
     * @minLength 1
     * @maxLength 28
     * @pattern ^[A-Za-zÀ-ÖØ-öø-ÿ0-9 .'&/()\-]+$
     */
  city: string;
  /**
     * X handle, with or without the leading @ (1-15 letters, digits or underscores)
     * @pattern ^@?[A-Za-z0-9_]{1,15}$
     */
  handle: string;
  /**
     * Role drawn (uppercased) before the chapter on the pill's second line, e.g. "Founder · INDIA CHAPTER". Omit for "Builder"; send an empty string for no role. Same alphabet as the city, at most 18 characters.
     * @maxLength 18
     * @pattern ^[A-Za-zÀ-ÖØ-öø-ÿ0-9 .'&/()\-]*$
     */
  role?: string;
  /** Optional profile photo as a data URL (PNG, JPEG or WebP, at most 25 MB decoded). Omit to render the neutral placeholder bust. */
  pfpDataUrl?: string;
  /**
     * Crop zoom, 1 (whole photo, cover-fit) to 4
     * @minimum 1
     * @maximum 4
     */
  scale?: number;
  /**
     * Horizontal pan as a fraction of the available slack, -1 (left edge) to 1 (right edge)
     * @minimum -1
     * @maximum 1
     */
  offsetX?: number;
  /**
     * Vertical pan as a fraction of the available slack, -1 (top edge) to 1 (bottom edge)
     * @minimum -1
     * @maximum 1
     */
  offsetY?: number;
}

export type BannerRenderResultMasterFormat = typeof BannerRenderResultMasterFormat[keyof typeof BannerRenderResultMasterFormat];


export const BannerRenderResultMasterFormat = {
  webp: 'webp',
} as const;

export interface BannerRenderResult {
  /** 1500x500 PNG, the file to upload as the X header */
  finalPngDataUrl: string;
  finalWidth: number;
  finalHeight: number;
  /** Size of the decoded PNG in bytes */
  finalBytes: number;
  /** 3000x1000 retina WebP (quality 95) */
  masterDataUrl: string;
  masterWidth: number;
  masterHeight: number;
  masterFormat: BannerRenderResultMasterFormat;
  /** Size of the decoded WebP in bytes */
  masterBytes: number;
  /** Suggested download file name without extension, e.g. arc-house-papercut-turkey-istanbul-vaibhav_0xq */
  fileStem: string;
  /** What the header reads, e.g. "ISTANBUL · @vaibhav_0xq" */
  chapterLine: string;
}