# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Arc House chapter banner API. Renders personalised papercut X headers server-side, pixel-identical to the in-browser preview. Nothing is persisted and photos are never stored or logged.
 * OpenAPI spec version: 0.1.0
 */
import type { ApiErrorField } from './apiErrorField';

export interface ApiError {
  /** Human-readable message safe to show inline in the studio */
  error: string;
  /** Which studio field the error belongs to, when it maps to one */
  field?: ApiErrorField;
}