# lib/api-zod/src/generated/types/bannerCountry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Arc House chapter banner API. Renders personalised papercut X headers server-side, pixel-identical to the in-browser preview. Nothing is persisted and photos are never stored or logged.
 * OpenAPI spec version: 0.1.0
 */

/**
 * Country chapter slug; selects the papercut base and paper palette
 */
export type BannerCountry = typeof BannerCountry[keyof typeof BannerCountry];


export const BannerCountry = {
  india: 'india',
  argentina: 'argentina',
  bangladesh: 'bangladesh',
  brazil: 'brazil',
  vietnam: 'vietnam',
  france: 'france',
  nigeria: 'nigeria',
  portugal: 'portugal',
  turkey: 'turkey',
  singapore: 'singapore',
} as const;