# lib/api-zod/src/generated/types/bannerRenderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Arc House chapter banner API. Renders personalised papercut X headers server-side, pixel-identical to the in-browser preview. Nothing is persisted and photos are never stored or logged.
 * OpenAPI spec version: 0.1.0
 */
import type { BannerCountry } from './bannerCountry';

export interface BannerRenderInput {
  country: BannerCountry;
  /**
     * City drawn (uppercased) under the chapter title. Letters, digits, spaces and . ' & / ( ) -
     * @minLength 1
     * @maxLength 28
     * @pattern ^[A-Za-zÀ-ÖØ-öø-ÿ0-9 .'&/()\-]+$
     */
  city: string;
  /**
     * X handle, with or without the leading @ (1-15 letters, digits or underscores)
     * @pattern ^@?[A-Za-z0-9_]{1,15}$
     */
  handle: string;
  /**
     * Role drawn (uppercased) before the chapter on the pill's second line, e.g. "Founder · INDIA CHAPTER". Omit for "Builder"; send an empty string for no role. Same alphabet as the city, at most 18 characters.
     * @maxLength 18
     * @pattern ^[A-Za-zÀ-ÖØ-öø-ÿ0-9 .'&/()\-]*$
     */
  role?: string;
  /** Optional profile photo as a data URL (PNG, JPEG or WebP, at most 25 MB decoded). Omit to render the neutral placeholder bust. */
  pfpDataUrl?: string;
  /**
     * Crop zoom, 1 (whole photo, cover-fit) to 4
     * @minimum 1
     * @maximum 4
     */
  scale?: number;
  /**
     * Horizontal pan as a fraction of the available slack, -1 (left edge) to 1 (right edge)
     * @minimum -1
     * @maximum 1
     */
  offsetX?: number;
  /**
     * Vertical pan as a fraction of the available slack, -1 (top edge) to 1 (bottom edge)
     * @minimum -1
     * @maximum 1
     */
  offsetY?: number;
}