# lib/api-zod/src/generated/types/bannerRenderResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Arc House chapter banner API. Renders personalised papercut X headers server-side, pixel-identical to the in-browser preview. Nothing is persisted and photos are never stored or logged.
 * OpenAPI spec version: 0.1.0
 */
import type { BannerRenderResultMasterFormat } from './bannerRenderResultMasterFormat';

export interface BannerRenderResult {
  /** 1500x500 PNG, the file to upload as the X header */
  finalPngDataUrl: string;
  finalWidth: number;
  finalHeight: number;
  /** Size of the decoded PNG in bytes */
  finalBytes: number;
  /** 3000x1000 retina WebP (quality 95) */
  masterDataUrl: string;
  masterWidth: number;
  masterHeight: number;
  masterFormat: BannerRenderResultMasterFormat;
  /** Size of the decoded WebP in bytes */
  masterBytes: number;
  /** Suggested download file name without extension, e.g. arc-house-papercut-turkey-istanbul-vaibhav_0xq */
  fileStem: string;
  /** What the header reads, e.g. "ISTANBUL · @vaibhav_0xq" */
  chapterLine: string;
}